Bugfix and header tweaks.
Fixed comp. view items.
Got rid of the integrated header because Sets, minor sizing tweaks.

The idea of the integrated header was to increase the size of the list by using the title bar, with Sets thats... well... a problem. It makes it worse by generating a lot of unused space.